Half way done with sewarchbar
I know I was going to start doing data analysis but I decided to add a better searchbar. At the moment, we have comp sector, subsector, industry, name, buyer name and buyer titles working, I need to add the two dates and the stock quantity, price (or maybe a joined thing).

I know it does not seem like a lot but because there is so much data it is actually quite nice being able to see and work with all of this on a better level. It will also be an importaint tool later